# Improved performance of removing comments from code